+++ lib/utils/handlers/item_sort.dart
@@ -2,7 +2,38 @@ import 'dart:io';
 import 'package:path/path.dart' as path;
 
 class ItemSortHandler {
-  static List<FileSystemEntity> sortItems(
+  static List<FileSystemEntity> getSortedItems(
+      List<FileSystemEntity> items, String folderPriority, String sortOrder) {
+    return _folderPrioritySort(items, folderPriority, sortOrder);
+  }
+
+  static List<FileSystemEntity> _folderPrioritySort(
+      List<FileSystemEntity> items, String folderPriority, String sortOrder) {
+    List<FileSystemEntity> folders = [];
+    List<FileSystemEntity> files = [];
+
+    for (var item in items) {
+      if (item is Directory) {
+        folders.add(item);
+      } else {
+        files.add(item);
+      }
+    }
+
+    folders = _sortItems(folders, sortOrder);
+    files = _sortItems(files, sortOrder);
+
+    switch (folderPriority) {
+      case 'above':
+        return [...folders, ...files];
+      case 'below':
+        return [...files, ...folders];
+      default:
+        return _sortItems(items, sortOrder);
+    }
+  }
+
+  static List<FileSystemEntity> _sortItems(
       List<FileSystemEntity> items, String sortOrder) {
     switch (sortOrder) {
       case 'default':